requirements.txt
fastapi
uvicorn
sqlalchemy
watchdog
requests
pypdf
faiss-cpu
streamlit
pandas
numpy 
types-requests
